docs/requirements.txt
# sphinx requirements
sphinxcontrib_django==2.5
sphinxcontrib_images==1.0.1
sphinxcontrib-plantuml==0.31

# requirements for the package... in case its dependencies aren't installed otherwise
Django==5.0.6
django-enumfield>=3.1
django-extensions==3.2.3
django-dramatiq==0.11.6
dramatiq[redis]==1.17.0
pydantic==2.12.5
requests==2.32.3


# requirements for the tests
pytest==8.2.2
pytest-django==4.8.0
